\name{bcVoc}
\alias{bcVoc}
\alias{pal}
\description{ provisional graph of topics associated with bioc 1.6 }
\non_function{}
\title{ provisional graph of topics associated with bioconductor 1.6  }
\usage{data(bcVoc); data(pal)}
\format{
  The format is:

Formal class 'graphNEL' [package "graph"] with 3 slots

  ..@ nodes   : chr [1:68] "vocRoot" "StatisticalModelingForHighThroughputBiology" "SequenceAnalysis" "SequenceDatabases" ...

  ..@ edgeL   :List of 68

  .. ..$ vocRoot                                    :List of 1

  .. .. ..$ edges: int [1:15] 6 8 9 10 2 42 68 50 3 5 ...

  .. ..$ StatisticalModelingForHighThroughputBiology:List of 1

  .. .. ..$ edges: int [1:11] 19 20 23 28 29 33 34 35 36 37 ...

}
\source{
demonstration vocabulary graph supplied with package.

pal is a demonstration version of a package - vocabulary association list.
}
\references{
None 
}
\keyword{models}
