\name{extractNEWS}
\alias{extractNEWS}
\title{Extract NEWS files from source package tarballs}
\description{
  Extracts NEWS files from source tarballs of packages.
}
\usage{
extractNEWS(reposRoot, srcContrib, destDir)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{reposRoot}{Top level path for CRAN-style repos}
  \item{srcContrib}{Location of source packages}
  \item{destDir}{where to extract}
}

\keyword{utilities}
