\name{extractReadmes}
\alias{extractReadmes}
\title{Extract README files from source package tarballs}
\description{
  Extracts README files from source tarballs of packages.
}
\usage{
extractReadmes(reposRoot, srcContrib, destDir)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{reposRoot}{Top level path for CRAN-style repos}
  \item{srcContrib}{Location of source packages}
  \item{destDir}{where to extract}
}

\keyword{utilities}
