\name{extractTopLevelFiles}
\alias{extractTopLevelFiles}
\title{Extract files from the top level of source package tarballs}
\description{
  Extracts files from source tarballs of packages.
}
\usage{
extractTopLevelFiles(reposRoot, srcContrib, destDir, fileName)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{reposRoot}{Top level path for CRAN-style repos}
  \item{srcContrib}{Location of source packages}
  \item{destDir}{where to extract}
  \item{fileName}{name of file to extract}
}

\keyword{utilities}
