\name{Htmlized-class}
\docType{class}
\alias{Htmlized-class}
\alias{htmlDoc,Htmlized-method}

\title{Class "Htmlized"}
\description{A virtual class for HTML serialization method dispatch.
}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Methods}{
  \describe{
    \item{htmlDoc}{\code{signature(object = "Htmlized")}: Return the
    html-ized representation of \code{object} as a complete HTML document.}
	 }
}
\author{Seth Falcon}
\keyword{classes}
