\name{htmlValue}

\alias{htmlValue}

\title{HTML Representation of an Object}
\description{
  This generic function should return an \code{XMLNode} instance
  representing the specified object in HTML
}
\usage{
htmlValue(object)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{An object}
}
\value{
  An instance of \code{XMLNode} from the \code{XML} package.
}
\author{Seth Falcon}
\seealso{\code{\link{htmlDoc}}, \code{\link{htmlFilename}}}
\keyword{methods}
