\name{writeHtmlDoc}
\alias{writeHtmlDoc}

\title{Write an XML DOM containing HTML to a file}
\description{
  Given a DOM tree from the XML package and a filename, write the DOM to
  disk creating an HTML file.
}
\usage{
writeHtmlDoc(html, file)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{html}{A DOM object from the XML package}
  \item{file}{A string giving the filename}
}
\author{S. Falcon}
\keyword{ utilities }
