% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_html_mans.R
\name{html-manuals}
\alias{html-manuals}
\alias{build_db_from_source}
\alias{build_meta_aliases_db}
\alias{build_meta_rdxrefs_db}
\alias{build_html_mans}
\title{Helper functions to generate HTML pages for reference manuals}
\usage{
build_db_from_source(package_dir, src_base)

build_meta_aliases_db(web_dir, aliases_db_file, force = FALSE)

build_meta_rdxrefs_db(web_dir, rdxrefs_db_file, force = FALSE)

build_html_mans(package_dirs, src_base)
}
\arguments{
\item{package_dir}{`character(1)` The local path to a package for which to
build the aliases and cross-ref databases (`rds` files).}

\item{src_base}{`character(1)` The path to the base hosting directory for the
package repository. This is typically a location on the BBS server.}

\item{web_dir}{`character(1)` The `web/packages` local directory that is
also hosted on the website e.g., for CRAN
\url{https://cran.r-project.org/web/packages/}}

\item{aliases_db_file}{`character(1)` The file path to `aliases.rds` file
generated by the `build_db_from_source` function.}

\item{force}{`logical(1)` If `FALSE`, the function will only update the
database entries for which the aliases/rdxrefs file is more recent than the
database file. If `TRUE`, the function will read all aliases/rdxrefs files.}

\item{rdxrefs_db_file}{`character(1)` The file path to `rdxrefs.rds` file
generated by the `build_db_from_source` function.}

\item{package_dirs}{`character()` A vector of local package directories.}
}
\description{
Helper functions to generate HTML pages for reference manuals
}
\examples{
if (interactive()) {
    library(BiocPkgTools)
    bioc_sub <- pkgBiocDeps(
        "SummarizedExperiment", pkgType = "software",
        recursive = TRUE, only.bioc = TRUE
    )
    bioc_sub <- unlist(bioc_sub, use.names = FALSE)

    ## generate from Bioc package source dirs
    packages <- file.path(normalizePath("~/bioc"), bioc_sub)
    src_base <- "~/minibioc/packages/3.20/bioc"

    for (package in packages) {
       build_db_from_source(package, src_base)
    }
}
if (interactive()) {
    src_base <- "~/minibioc/packages/3.20/bioc/"
    web_dir <- file.path(src_base, "web", "packages")

    meta_folder <- file.path(contrib.url(src_base), "Meta")
    if (!dir.exists(meta_folder)) dir.create(meta_folder, recursive = TRUE)
    aliases_db_file <- file.path(meta_folder, "aliases.rds")

    meta_aliases_db <- build_meta_aliases_db(web_dir, aliases_db_file)

    saveRDS(meta_aliases_db, aliases_db_file, version = 2)
}
if (interactive()) {
    src_base <- "~/minibioc/packages/3.20/bioc/"
    web_dir <- file.path(src_base, "web", "packages")

    meta_folder <- file.path(contrib.url(src_base), "Meta")
    if (!dir.exists(meta_folder)) dir.create(meta_folder, recursive = TRUE)
    rdxrefs_db_file <- file.path(meta_folder, "rdxrefs.rds")

    meta_rdxrefs_db <- build_meta_rdxrefs_db(web_dir, rdxrefs_db_file)

    saveRDS(meta_rdxrefs_db, rdxrefs_db_file, version = 2)
}
if (interative()) {
    library(BiocPkgTools)
    bioc_sub <- pkgBiocDeps(
        "SummarizedExperiment", pkgType = "software",
        recursive = TRUE, only.bioc = TRUE
    )
    bioc_sub <- unlist(bioc_sub, use.names = FALSE)

    ## generate from Bioc package source dirs
    packages <- file.path(normalizePath("~/bioc"), bioc_sub)
    src_base <- "~/minibioc/packages/3.20/bioc"

    build_html_mans(packages, src_base)
}
}
